Change for adapte written convention

+++ Assets/_Core/Event/ExoEvent/Scripts/ScriptableEventInt.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "new_" + nameof(ScriptableEventInt), menuName = "ExoEvent/ScriptableEventInt")]
 public class ScriptableEventInt : ScriptableObject
 {
-    public Action<int> Event;
+    public Action<float> Event;
 }
